add framer motion + minor changes
