+++ app/containers/gallery/Header.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const Header = () => {
+  return (
+    <>
+      <h1 className='w-full py-12 px-6 font-bold text-4xl md:text-5xl lg:text-7xl tracking-[.05em] lg:leading-[5rem] text-center'>
+        Some of Our Favorite{' '}
+        <span className='text-[#B4794D] font-wf text-6xl md:text-7xl lg:text-8xl'>
+          Mementos
+        </span>
+      </h1>
+    </>
+  );
+};
+
+export default Header;
